refactor: replace generate_toy with generate_toy2
