+++ tf_pwa/generator/__init__.py
@@ -1,1 +1,1 @@
-from .linear_interpolation import GenTest
+from .generator import BaseGenerator, GenTest
